Added small patterns to activities. Corrected histograms for sizes in colorfunction 10/11 : NB not the right place to do histograms since incomplete data! Checked code with Xcode analyze and made a number of small corrections. Revised article abstract to reflect recent developments.

+++ fastgenegol/genelifec/maxmatch.c
@@ -123,7 +123,7 @@ boolean bfs(int m, short unsigned int kk[], unsigned int ii[], short unsigned in
                 {
                     // Consider the pair and add it to queue
                     dist[pairV[v]] = dist[u] + 1;
-                    queue_push(pairV[v]);
+                    queue_push((int) pairV[v]);
                 }
             }
         }
